CrelteRequest #36
`CrelteRouted` becomes `CrelteRequest`, which has a `req: Request` instead of the Route
this allows to specify special things which only concern routes which are not completed
It is now possible to remove the onRequest opts argument because it can be in the
Request, this also allows loadData calls to modify the current request.

This removes all stuff related to NextRoute and NextSite. use router.onRequest if needed

add `GraphQl.request` && make GraphQlQuery an interface