core: add `beforeQueryEntry`, `queryEntry` and `afterQueryEntry` events - this adds three events all useful for similar things
This event is useful if you want to do some sort of aliasing or want to add
an addition entry variable (passed to gql). An example is a filter with should be stored
in an uri segment. You can rewrite the uri that is passed to entry.graphql allowing
to hide the last url segment for craft.

This event is the most powerful because you can override the
entire entry.graphql call. Calls are nested so the last event handler which
gets added can decide if it want's to call all previous handler.
Use this for example if you want to add retour support if the page was not found.

This event is intended to modify the entry object after it was queried.
For example if some property is missing from the object, like localized on
commerce products, you can run you're own query to check the other sites.
Or if you want to override a section handle or similar.